+++ backend/canisters/group/impl/src/lib.rs
@@ -445,7 +445,7 @@ struct Data {
     pub pending_payments_queue: PendingPaymentsQueue,
     pub total_payment_receipts: PaymentReceipts,
     pub video_call_operators: Vec<Principal>,
-    #[serde(default = "ic_root_key")]
+    #[serde(with = "serde_bytes", default = "ic_root_key")]
     pub ic_root_key: Vec<u8>,
     pub event_store_client: EventStoreClient<CdkRuntime>,
 }