+++ backend/integration_tests/src/lib.rs
@@ -3,7 +3,6 @@
 use crate::utils::principal_to_username;
 use candid::Principal;
 use pocket_ic::PocketIc;
-use serde_bytes::ByteBuf;
 use types::{CanisterId, Cycles, UserId};
 
 mod batched_summary_and_event_tests;
@@ -60,7 +59,7 @@ pub struct TestEnv {
 pub struct User {
     pub principal: Principal,
     pub user_id: UserId,
-    pub public_key: ByteBuf,
+    pub public_key: Vec<u8>,
 }
 
 impl User {