+++ backend/libraries/testing/src/rng.rs
@@ -1,7 +1,6 @@
 use crate::NNS_INTERNET_IDENTITY_CANISTER_ID;
 use candid::Principal;
 use rand::{random, RngCore};
-use serde_bytes::ByteBuf;
 use types::{CanisterId, MessageId};
 
 pub fn random_principal() -> Principal {
@@ -10,11 +9,11 @@ pub fn random_principal() -> Principal {
     Principal::from_slice(&random_bytes)
 }
 
-pub fn random_internet_identity_principal() -> (Principal, ByteBuf) {
+pub fn random_internet_identity_principal() -> (Principal, Vec<u8>) {
     random_delegated_principal(NNS_INTERNET_IDENTITY_CANISTER_ID)
 }
 
-pub fn random_delegated_principal(originating_canister_id: CanisterId) -> (Principal, ByteBuf) {
+pub fn random_delegated_principal(originating_canister_id: CanisterId) -> (Principal, Vec<u8>) {
     let algorithm_bytes = [48u8, 60, 48, 12, 6, 10, 43, 6, 1, 4, 1, 131, 184, 67, 1, 2, 3, 44, 0];
     let random_bytes: [u8; 32] = random();
 
@@ -23,7 +22,7 @@ pub fn random_delegated_principal(originating_canister_id: CanisterId) -> (Princ
     public_key.extend_from_slice(originating_canister_id.as_slice());
     public_key.extend_from_slice(&random_bytes);
 
-    (Principal::self_authenticating(&public_key), ByteBuf::from(public_key))
+    (Principal::self_authenticating(&public_key), public_key)
 }
 
 pub fn random_string() -> String {