+++ backend/canisters/identity/CHANGELOG.md
@@ -1,9 +1,17 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
+
+### Changed
+
+- Remove `principal` from `create_identity` response ([#5849](https://github.com/open-chat-labs/open-chat/pull/5849))
+- Expose count of auth principals per originating canister in metrics ([#5851](https://github.com/open-chat-labs/open-chat/pull/5851))
+
+## [[2.0.1177](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1177-identity)] - 2024-05-23
 
 ### Changed